+++ src/xenia/kernel/xam/xam_msg.cc
@@ -25,7 +25,8 @@ dword_result_t XMsgInProcessCall_entry(dword_t app, dword_t message,
   auto result = kernel_state()->app_manager()->DispatchMessageSync(app, message,
                                                                    arg1, arg2);
   if (result == X_ERROR_NOT_FOUND) {
-    XELOGE("XMsgInProcessCall: app {:08X} undefined", app);
+    XELOGE("XMsgInProcessCall: app {:08X} undefined",
+           static_cast<uint32_t>(app));
   }
   return result;
 }
@@ -37,7 +38,8 @@ dword_result_t XMsgSystemProcessCall_entry(dword_t app, dword_t message,
   auto result = kernel_state()->app_manager()->DispatchMessageAsync(
       app, message, buffer, buffer_length);
   if (result == X_ERROR_NOT_FOUND) {
-    XELOGE("XMsgSystemProcessCall: app {:08X} undefined", app);
+    XELOGE("XMsgSystemProcessCall: app {:08X} undefined",
+           static_cast<uint32_t>(app));
   }
   return result;
 }
