+++ src/xenia/kernel/xam/xam_user.cc
@@ -412,7 +412,7 @@ dword_result_t XamUserWriteProfileSettings_entry(
         "XamUserWriteProfileSettings: setting index [{}]:"
         " from={} setting_id={:08X} data.type={}",
         n, (uint32_t)setting.from, (uint32_t)setting.setting_id,
-        setting.data.type);
+        static_cast<uint32_t>(setting.data.type));
 
     switch (setting_type) {
       case X_USER_DATA_TYPE::CONTENT:
@@ -445,7 +445,7 @@ dword_result_t XamUserWriteProfileSettings_entry(
       case X_USER_DATA_TYPE::DATETIME:
       default: {
         XELOGE("XamUserWriteProfileSettings: Unimplemented data type {}",
-               setting_type);
+               static_cast<uint32_t>(setting_type));
       } break;
     };
   }
